4.2. Результат работы метода [#232622]

+++ src/ru/job4j/calculator/MathFunc.java
@@ -1,19 +1,19 @@
 package ru.job4j.calculator;
 
 public class MathFunc {
-    public static int func1(int x) {
+    public static float func1(float x) {
         return x * x + 1;
     }
 
-    public static int func2(int x) {
+    public static float func2(float x) {
         return 1 / x;
     }
 
     public static void main(String[] args) {
-        int result1 = MathFunc.func1(3);
-        int result2 = MathFunc.func2(5);
-        int result3 = MathFunc.func1(100);
-        int total = result1 + result2;
+        float result1 = MathFunc.func1(3);
+        float result2 = MathFunc.func2(5);
+        float result3 = MathFunc.func1(100);
+        float total = result1 + result2;
         System.out.println(total);
         System.out.println(result3);
     }
